ウィンドウ幅などの修正

+++ src/emg.cpp
@@ -7,10 +7,10 @@ volatile float extensor_value = 0;
 volatile float flexor_value = 0;
 
 volatile int begin_index = 0;
-volatile float raw_extensor_values[RAW_EMG_LENGTH] = {0};
-volatile float raw_flexor_values[RAW_EMG_LENGTH] = {0};
-volatile float ar_extensor_values[RAW_EMG_LENGTH] = {0};
-volatile float ar_flexor_values[RAW_EMG_LENGTH] = {0};
+volatile int raw_extensor_values[RAW_EMG_LENGTH] = {0};
+volatile int raw_flexor_values[RAW_EMG_LENGTH] = {0};
+volatile int ar_extensor_values[RAW_EMG_LENGTH] = {0};
+volatile int ar_flexor_values[RAW_EMG_LENGTH] = {0};
 volatile float n_extensor_values[RAW_EMG_LENGTH] = {0};
 volatile float n_flexor_values[RAW_EMG_LENGTH] = {0};
 volatile float ra_extensor_values[RAW_EMG_LENGTH] = {0};
